fix: reindex method generation, add option for disable reindex related
fixes a bug where the reindex method generation would incorrectly
determine the first index in a range, because we were ranging twice and
only checking the inside index value. this now adds the two indexes
together to determine if it's in the first iteration.

adds an option for disabling reindex related methods. this is useful for
preventing a bunch of unused code, in cases where nested objects should
never trigger a reindex.