feat: add onSuccess and onFailure parameters to bulk index methods
The BulkIndexSync function did not take onSuccess or onFailure functions
as parameters before. So it wasn't possible to add checking to ensure
that a new elasticsearch bulk indexer completed indexing all of the
documents provided to it. The only errors that it would return are
errors when adding documents or closing and flushing the indexer.

these new parameters on the bulk model's methods should allow for
checking if all documents succeeded or not.